feat!: Add double strike and strike cooldown

+++ src/player/strike.rs
@@ -1,27 +1,51 @@
+use std::time::Duration;
+
 use bevy::prelude::*;
 use bevy_rapier2d::prelude::*;
 use bevy_trickfilm::prelude::*;
 
-use crate::{utils::quat_from_vec2, world::camera::YSort, GameAssets, GameState};
+use crate::{
+    utils::{quat_from_vec2, FixedRotation},
+    world::camera::YSort,
+    GameAssets, GameState,
+};
 
 use super::{
     input::{MouseWorldCoords, PlayerInput},
     Player,
 };
 
+const OFFSET: Vec3 = Vec3::new(0.0, -10.0, 0.0);
+const CHAIN_COOLDOWN: f32 = 0.5;
+const CHAIN_STRIKE_COOLDOWN: f32 = 0.25;
+const STRIKE_COOLDOWN: f32 = 0.5;
+
+#[derive(Resource, Default)]
+struct StrikeCooldown {
+    absolute_cooldown: Timer,
+    chain_cooldown: Timer,
+}
+
 #[derive(Component)]
 struct Strike;
 
 #[derive(Event)]
 pub struct SpawnStrike {
-    pub transform: Transform,
+    pub rot: Quat,
+    strike_index: usize,
 }
 
 fn spawn_strikes(
     mut commands: Commands,
     assets: Res<GameAssets>,
+    q_player: Query<Entity, With<Player>>,
     mut ev_spawn_strike: EventReader<SpawnStrike>,
 ) {
+    let player_entity = match q_player.get_single() {
+        Ok(r) => r,
+        Err(_) => return,
+    };
+
     for ev in ev_spawn_strike.read() {
         let mut animator = AnimationPlayer2D::default();
         animator.play(assets.player_strike_animations[0].clone());
@@ -37,18 +61,28 @@ fn spawn_strikes(
             ))
             .id();
 
-        commands
+        let strike = commands
             .spawn((
+                FixedRotation {
+                    offset: OFFSET,
+                    rot: ev.rot,
+                },
                 Strike,
                 YSort(1.0),
                 animator,
                 SpriteSheetBundle {
                     texture_atlas: assets.player_strike.clone(),
-                    transform: ev.transform.clone(),
+                    sprite: TextureAtlasSprite {
+                        flip_y: ev.strike_index != 0,
+                        ..default()
+                    },
                     ..default()
                 },
             ))
-            .push_children(&[collider]);
+            .push_children(&[collider])
+            .id();
+
+        commands.entity(player_entity).push_children(&[strike]);
     }
 }
 
@@ -64,9 +98,10 @@ fn despawn_strikes(
 }
 
 fn trigger_strike(
-    q_player: Query<&Transform, With<Player>>,
     player_input: Res<PlayerInput>,
     mouse_coords: Res<MouseWorldCoords>,
+    mut strike_cooldown: ResMut<StrikeCooldown>,
+    q_player: Query<&Transform, With<Player>>,
     mut ev_spawn_strike: EventWriter<SpawnStrike>,
 ) {
     let player_transform = match q_player.get_single() {
@@ -74,28 +109,53 @@ fn trigger_strike(
         Err(_) => return,
     };
 
+    if !strike_cooldown.absolute_cooldown.finished() {
+        return;
+    }
+
     let rot = quat_from_vec2(mouse_coords.0 - player_transform.translation.truncate());
 
+    if strike_cooldown.chain_cooldown.just_finished() {
+        strike_cooldown
+            .absolute_cooldown
+            .set_duration(Duration::from_secs_f32(STRIKE_COOLDOWN));
+        strike_cooldown.absolute_cooldown.reset();
+    }
+
     if player_input.attack {
-        let transform = Transform::from_translation(player_transform.translation)
-            .with_rotation(rot)
-            .with_scale(player_transform.scale);
-        ev_spawn_strike.send(SpawnStrike { transform })
+        // This is the first strike
+        if strike_cooldown.chain_cooldown.finished() {
+            strike_cooldown
+                .chain_cooldown
+                .set_duration(Duration::from_secs_f32(CHAIN_COOLDOWN));
+            strike_cooldown.chain_cooldown.reset();
+
+            strike_cooldown
+                .absolute_cooldown
+                .set_duration(Duration::from_secs_f32(CHAIN_STRIKE_COOLDOWN));
+            strike_cooldown.absolute_cooldown.reset();
+
+            ev_spawn_strike.send(SpawnStrike {
+                rot,
+                strike_index: 0,
+            });
+        } else {
+            strike_cooldown
+                .absolute_cooldown
+                .set_duration(Duration::from_secs_f32(STRIKE_COOLDOWN));
+            strike_cooldown.absolute_cooldown.reset();
+
+            ev_spawn_strike.send(SpawnStrike {
+                rot,
+                strike_index: 1,
+            });
+        }
     }
 }
 
-fn move_strikes(
-    q_player: Query<&Transform, With<Player>>,
-    mut q_strikes: Query<&mut Transform, (With<Strike>, Without<Player>)>,
-) {
-    let player_transform = match q_player.get_single() {
-        Ok(r) => r,
-        Err(_) => return,
-    };
-
-    for mut transform in &mut q_strikes {
-        transform.translation = player_transform.translation;
-    }
+fn tick_strike_cooldown(time: Res<Time>, mut strike_cooldown: ResMut<StrikeCooldown>) {
+    strike_cooldown.chain_cooldown.tick(time.delta());
+    strike_cooldown.absolute_cooldown.tick(time.delta());
 }
 
 pub struct PlayerStrikePlugin;
@@ -104,9 +164,15 @@ impl Plugin for PlayerStrikePlugin {
     fn build(&self, app: &mut App) {
         app.add_systems(
             Update,
-            (spawn_strikes, despawn_strikes, trigger_strike, move_strikes)
+            (
+                spawn_strikes,
+                despawn_strikes,
+                trigger_strike,
+                tick_strike_cooldown,
+            )
                 .run_if(in_state(GameState::Gaming)),
         )
+        .init_resource::<StrikeCooldown>()
         .add_event::<SpawnStrike>();
     }
 }
